# Optimize T-Digest structure for small distributions