Dynamically increase the size of underlying arrays as values are added to digest.

Benchmark results comparing before/after optimization are displayed below:

```
BenchmarkTDigest.benchmarkCopyBEFORE		N/A	avgt	30	3328.574 ±  79.584 ns/op
BenchmarkTDigest.benchmarkCopyAFTER		N/A	avgt	30	1969.757 ±  19.348 ns/op
BenchmarkTDigest.benchmarkDeserializeBEFORE	N/A	avgt	30	1623.594 ± 111.647 ns/op
BenchmarkTDigest.benchmarkDeserializeAFTER	N/A	avgt	30	 165.251 ±   5.514 ns/op
BenchmarkTDigest.benchmarkInsertsBEFORE		N/A	avgt	30	  67.205 ±   0.796 ns/op
BenchmarkTDigest.benchmarkInsertsAFTER		N/A	avgt	30	  69.787 ±   2.767 ns/op
BenchmarkTDigest.benchmarkMergeBEFORE		N/A	avgt	30	4736.826 ±  24.052 ns/op
BenchmarkTDigest.benchmarkMergeAFTER		N/A	avgt	30	2181.707 ±  49.311 ns/op
BenchmarkTDigest.benchmarkSerializeBEFORE	N/A	avgt	30	2423.662 ± 138.643 ns/op
BenchmarkTDigest.benchmarkSerializeAFTER	N/A	avgt	30	 133.829 ±   4.896 ns/op
```

```
== NO RELEASE NOTE ==
```
Replaces #13160